backend/backend/requirements.txt
Django==5.2.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
mysqlclient==2.2.7
django-cors-headers==4.7.0
python-dotenv==1.1.1
django-filter